+++ src/transformer/events/mod_feedback/response_submitted/response_submitted.php
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU General Public License
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 
-namespace src\transformer\events\mod_feedback;
+namespace src\transformer\events\mod_feedback\response_submitted;
 
 defined('MOODLE_INTERNAL') || die();
