+++ tests/mod_feedback/response_submitted/no_items/data.json
@@ -24,5 +24,6 @@
             "id": 1,
             "name": "test_feedback_name"
         }
-    ]
+    ],
+    "feedback_value": []
 }
