+++ tests/mod_feedback/response_submitted/no_items/statements.json
@@ -31,7 +31,7 @@
                     "http:\/\/moodle.org": "1.0.0",
                     "https:\/\/github.com\/xAPI-vle\/moodle-logstore_xapi": "0.0.0-development",
                     "event_name": "\\mod_feedback\\event\\response_submitted",
-                    "event_function": "\\src\\transformer\\events\\mod_feedback\\response_submitted"
+                    "event_function": "\\src\\transformer\\events\\mod_feedback\\response_submitted\\handler"
                 }
             },
             "contextActivities": {
