+++ tests/mod_feedback/response_submitted/unknown_typ/data.json
@@ -1,0 +1,43 @@
+{
+    "user": [
+        {
+            "id": 1,
+            "firstname": "test_fullname",
+            "email": "test@test.com"
+        }
+    ],
+    "course": [
+        {
+            "id": 1,
+            "fullname": "test_name",
+            "lang": "en"
+        }
+    ],
+    "feedback_completed": [
+        {
+            "id": 1,
+            "feedback": 1
+        }
+    ],
+    "feedback": [
+        {
+            "id": 1,
+            "name": "test_feedback_name"
+        }
+    ],
+    "feedback_item": [
+        {
+            "id": 1,
+            "feedback": 1,
+            "name": "test_feedback_item",
+            "typ": "unknown"
+        }
+    ],
+    "feedback_value": [
+        {
+            "id": 1,
+            "item": 1,
+            "completed": 1
+        }
+    ]
+}
