+++ tests/mod_feedback/response_submitted/unknown_typ/event.json
@@ -1,0 +1,9 @@
+{
+    "id": 1,
+    "userid": 1,
+    "courseid": 1,
+    "timecreated": 1433946701,
+    "objectid": 1,
+    "contextinstanceid": 1,
+    "eventname": "\\mod_feedback\\event\\response_submitted"
+}
